src/requirements.txt
asgiref==3.3.1
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==2.2
django-filter==2.4.0
django-guardian==2.3.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
idna==2.10
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.17
psycopg2-binary==2.8.6
ptyprocess==0.7.0
Pygments==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
simplejson==3.17.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
wcwidth==0.2.5